polish DeleteModal

+++ src/app/components/delete/DeleteModal.js
@@ -6,7 +6,7 @@ import { useApi, useNotifications, useEventManager, useLoading, Alert, ErrorButt
 import { clearContacts, deleteContacts } from 'proton-shared/lib/api/contacts';
 import { allSucceded } from 'proton-shared/lib/api/helpers/response';
 
-const DeleteModal = ({ beDeletedIDs, deleteAll, onDelete, ...rest }) => {
+const DeleteModal = ({ contactIDs = [], deleteAll, onDelete, ...rest }) => {
     const api = useApi();
     const { createNotification } = useNotifications();
     const { call } = useEventManager();
@@ -17,20 +17,20 @@ const DeleteModal = ({ beDeletedIDs, deleteAll, onDelete, ...rest }) => {
     const handleDelete = async () => {
         if (deleteAll) {
             await api(clearContacts());
-            onDelete();
+            onDelete && onDelete();
             await call();
             rest.onClose();
             return createNotification({ text: c('Success').t`Contacts deleted` });
         }
-        const apiSuccess = allSucceded(await api(deleteContacts(beDeletedIDs)));
-        onDelete();
+        const apiSuccess = allSucceded(await api(deleteContacts(contactIDs)));
+        onDelete && onDelete();
         await call();
         rest.onClose();
         if (!apiSuccess) {
             return createNotification({ text: c('Error').t`Some contacts could not be deleted`, type: 'warning' });
         }
         createNotification({
-            text: c('Success').ngettext(msgid`Contact deleted`, `Contacts deleted`, beDeletedIDs.length)
+            text: c('Success').ngettext(msgid`Contact deleted`, `Contacts deleted`, contactIDs.length)
         });
     };
     return (
@@ -45,7 +45,7 @@ const DeleteModal = ({ beDeletedIDs, deleteAll, onDelete, ...rest }) => {
                 {c('Warning').ngettext(
                     msgid`This action will permanently delete the selected contact. Are you sure you want to delete this contact?`,
                     `This action will permanently delete selected contacts. Are you sure you want to delete these contacts?`,
-                    beDeletedIDs.length
+                    contactIDs.length
                 )}
             </Alert>
         </FormModal>
@@ -53,7 +53,7 @@ const DeleteModal = ({ beDeletedIDs, deleteAll, onDelete, ...rest }) => {
 };
 
 DeleteModal.propTypes = {
-    beDeletedIDs: PropTypes.arrayOf(PropTypes.string),
+    contactIDs: PropTypes.arrayOf(PropTypes.string),
     deleteAll: PropTypes.bool,
     onDelete: PropTypes.func
 };
